TradeBot/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==2.6.1
asgiref==3.4.1
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
binance==0.3
bta-lib==1.0.0
cached-property==1.5.2
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.11.0
dateparser==1.1.0
DateTime==4.3
Deprecated==1.2.13
Django==3.0.6
django-background-tasks==1.2.5
django-celery-beat==2.2.1
django-ckeditor==5.9.0
django-compat==1.0.15
django-cors-headers==3.2.1
django-extensions==2.2.9
django-js-asset==1.2.2
django-timezone-field==4.2.3
fonttools==4.28.5
frozenlist==1.2.0
idna==3.3
importlib-metadata==4.10.0
kiwisolver==1.3.2
kombu==4.6.11
matplotlib==3.5.1
multidict==5.2.0
numpy==1.21.4
packaging==21.3
pandas==1.3.5
Pillow==7.1.2
prompt-toolkit==3.0.24
pyparsing==3.0.6
python-binance==1.0.15
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.1
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.3.1
TA-Lib==0.4.24
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
Unidecode==1.1.1
urllib3==1.26.7
vine==1.3.0
wcwidth==0.2.5
websockets==9.1
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
zope.interface==5.4.0
